Draft 2 - using a for loop
I used a for loop for this part and created a list of the lanes the drunk person visits. I think this method will be worthwhile to consider as it may save some space on the code. I am also trying to link this to the probability code as Alisha and I will try use the method you've attempted. I used the open function to open the csv file of the probabilities and I aim to access the corresponding brackets of probabilities and the corresponding probabilities within those using an array or a list.