Update Attempt 6.py
Initially the random.choices method on line 68 was not reading the probabilities (in the weights part) because the function would output them as a string. I made changes to output them as floats. I also added the step variable from line 36 onward. This tells us the direction in which the first step was added. Lastly I removed the first if statement in the for loop. Please let me know if it is okay.