Updated Attempt 6
The probabilities weren't being read correctly in the previous code, hence the code was still reaching 0 and we were getting an error when it reached lane 10. I added a 'strip' function within the function1 code so the brackets and commas can be deleted from the probabilities list and added a for loop so the correct probabilities can be added to the list as a float. This allowed the 'weights' part to read and apply the correct probabilities. 

Also added a for loop to simulate the generation of steps.

Also added a frequency count for each lane within a for loop with increasing increments of 10. But this only does it for each separate simulation, not the total of the entire simulation.