Corrected TicTacToe
Previously there was an error when you input a lower case letter but it would still add a cross to the board. If you repeat the same lowercase input it says it's already taken but also shows that the input is invalid which isn't right. I've made some changes and this issue should be resolved. Please check to see if I've missed anything.