Lane frequencies
Instead of printing the respective frequencies individually, I created lists for each lane that we can store these in. This will be useful when we are trying to find the frequency that the person ended up in each lane after 10, 20, ..., 90, 100 steps, as we are able to sum the elements in the lists for all simulations. We should try figure out how to group them according to the number of steps now, and output these.