Update practiceq1.py
I changed lines 35, 45, 117 and 127 from ".lower" into ".upper" so that if the user enters in lower case it will automatically output the upper case. In line 142 and 145, I swapped "O" with "X". Previously, if player 1 was O and X won the game, then this would denote to player 1 winning which is wrong. This change allows the correct player to print when they win.  On lines 110 and 194, I changed the 9 into a 10 because previously if someone won after 9 moves it would give it as a tie, but now it doesn't give this issue.