Update Attempt 6.py
I've added a for loop to run through each of the steps and record each consecutive sequence. The code then outputs the longest consecutive sequence for each of the directions. At the moment, it only outputs this information for each simulation but not the the 1000 simulations as a  whole. I'll work on this next.