Update practiceq1.py
I've added a few empty line print statements before and after each player is asked to enter their turn so that the display in the python shell isn't as squashed. I've also added a few extra comments. Please confirm you're happy with the layout of the empty lines when you run the code.